feat: created unbounded napmap
